// generated/src/app/api-models/cap-model-rate-search.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class CapModelRateSearch {
  capModelId: string;
  rateSearchType: string;
  rateSearchOrder: number;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}